tues even

+++ src/java/DatabaseInserter.java
@@ -45,7 +45,7 @@ public class DatabaseInserter {
     this.dmf = dbm.getDialogMainFrame();
     require.invoke(Clojure.read("ln.codax-manager"));
     	 IFn getSessionID = Clojure.var("ln.codax-manager", "get-session-id");
-	 session_id = ((Long)getSessionID.invoke()).intValue();
+	 session_id = (int)getSessionID.invoke();
     // this.utils = dmf.getUtilities();
     //this.session = dmf.getSession();
   }
